refactoring of helper classes
